Make :find_user into a standalone plug
